split user assets and swaps store out
